aw080 resync to trunk, win build working, no checks yet

+++ sc/source/ui/docshell/docsh.cxx
@@ -121,9 +121,7 @@
 
 #include "docsh.hxx"
 #include "docshimp.hxx"
-//IAccessibility2 Implementation 2009-----
 #include <sfx2/viewfrm.hxx>
-//-----IAccessibility2 Implementation 2009
 #include <rtl/logfile.hxx>
 
 #include <comphelper/processfactory.hxx>
@@ -206,7 +204,7 @@ namespace
         static const struct
         {
             const char * mpFilterName;
-            unsigned mnFilterNameLen;
+            size_t mnFilterNameLen;
         } szMSFilterNames [] =
         {
             { pFilterExcel4, strlen( pFilterExcel4 ) },
@@ -1595,7 +1593,6 @@ sal_Bool __EXPORT ScDocShell::SaveAs( SfxMedium& rMedium )
 
     PrepareSaveGuard aPrepareGuard( *this);
 
-//IAccessibility2 Implementation 2009-----
     aDocument.setDocAccTitle(String());
     // SfxViewFrame* pFrame1 = SfxViewFrame::GetFirst( this, TYPE(SfxTopViewFrame));
     SfxViewFrame* pFrame1 = SfxViewFrame::GetFirst( this );
@@ -1611,7 +1608,6 @@ sal_Bool __EXPORT ScDocShell::SaveAs( SfxMedium& rMedium )
             }
         }
     }
-//-----IAccessibility2 Implementation 2009
     //  wait cursor is handled with progress bar
     sal_Bool bRet = SfxObjectShell::SaveAs( rMedium );
     if( bRet )
